Rework all tests.
