+++ tests/Stubs/ActiveRecord/Order.php
@@ -194,14 +194,6 @@ final class Order extends ActiveRecord
         });
     }
 
-    public function attributeLabels(): array
-    {
-        return [
-            'customer_id' => 'Customer',
-            'total' => 'Invoice Total',
-        ];
-    }
-
     public function activeAttributes(): array
     {
         return [
