+++ tests/Stubs/ActiveRecord/OrderItem.php
@@ -60,7 +60,7 @@ final class OrderItem extends ActiveRecord
 
     public function getCustom(): ActiveQuery
     {
-        return (new Order($this->db))->find();
+        return new ActiveQuery(Order::class, $this->db);
     }
 
     public function setTableName(?string $value): void
